unit particles can now have float values for emission rates. This allows also emission rates smaller than 1 like 0.01 for example

+++ source/glest_game/graphics/particle_type.cpp
@@ -111,7 +111,7 @@ void ParticleSystemType::load(const XmlNode *particleSystemNode, const string &d
 
 	//emission rate
 	const XmlNode *emissionRateNode= particleSystemNode->getChild("emission-rate");
-	emissionRate= emissionRateNode->getAttribute("value")->getIntValue();
+	emissionRate= emissionRateNode->getAttribute("value")->getFloatValue();
 
 	//energy max
 	const XmlNode *energyMaxNode= particleSystemNode->getChild("energy-max");
@@ -234,7 +234,7 @@ void ParticleSystemTypeSplash::load(const string &dir, const string &path,Render
 
 		//emission rate fade
 		const XmlNode *emissionRateFadeNode= particleSystemNode->getChild("emission-rate-fade");
-		emissionRateFade= emissionRateFadeNode->getAttribute("value")->getIntValue();
+		emissionRateFade= emissionRateFadeNode->getAttribute("value")->getFloatValue();
 		
 		//spread values
 		const XmlNode *verticalSpreadNode= particleSystemNode->getChild("vertical-spread");
